Share gRPC reflection client construction with requests
